Finish refactoring project 7
