+++ proj9DouglasHanssenMacDonaldZhang/CodeAreaTabPane.java
@@ -53,59 +53,4 @@ public class CodeAreaTabPane extends TabPane {
             return null;
         }
     }
-
-    /*
-     * Check for any missing braces, brackets or parentheses
-     */
-    public void handleCheckWellFormed() {
-        CodeArea activeCodeArea = this.getCurCodeArea();
-        // get the text of the current codeArea
-        String text = activeCodeArea.getText();
-        // now go through the text to check if it is malformed
-        long nOpenBraces = text.chars().filter(ch -> ch == '{').count();
-        long nCloseBraces= text.chars().filter(ch -> ch == '}').count();
-        long nOpenParens = text.chars().filter(ch -> ch == '(').count();
-        long nCloseParens = text.chars().filter(ch -> ch == ')').count();
-        long nOpenBrackets = text.chars().filter(ch -> ch == '[').count();
-        long nCloseBrackets = text.chars().filter(ch -> ch == ']').count();
-        String bracesMessage;
-        String parensMessage;
-        String bracketsMessage;
-        // determine if brackets are well formed
-        if (nOpenBraces > nCloseBraces) {
-            bracesMessage = "Missing " + (nOpenBraces - nCloseBraces) + " close braces\n";
-        }
-        else if (nOpenBraces < nCloseBraces) {
-            bracesMessage = "Missing " + (nCloseBraces - nOpenBraces) + " open braces\n";
-        }
-        else {
-            bracesMessage = "Braces are well formed\n";
-        }
-         // determine if parenthesis are well formed
-        if (nOpenParens > nCloseParens) {
-            parensMessage = "Missing " + (nOpenParens - nCloseParens) + " close parenthesis\n";
-        }
-        else if (nOpenParens < nCloseParens) {
-            parensMessage = "Missing " + (nCloseParens - nOpenParens) + " open parenthesis\n";
-        }
-        else {
-            parensMessage = "Parenthesis are well formed\n";
-        }
-         // determine if brackets are well formed
-        if (nOpenBrackets > nCloseBrackets) {
-            bracketsMessage = "Missing " + (nOpenBrackets - nCloseBrackets) + " close brackets\n";
-        }
-        else if (nOpenBrackets < nCloseBrackets) {
-            bracketsMessage = "Missing " + (nCloseBrackets - nOpenBrackets) + " open brackets\n";
-        }
-        else {
-            bracketsMessage = "Brackets are well formed\n";
-        }
-         // show messages
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-        alert.setTitle("Code formation report");
-        alert.setHeaderText("Checking brackets, parentheses and braces");
-        alert.setContentText(bracesMessage + parensMessage + bracketsMessage);
-        alert.showAndWait();
-    }
 }
